+++ changelog.md
@@ -2,7 +2,7 @@
 1. graphql: add graphiql
 2. add route: suport graphql
 3. lifespan: event_type support `shutdown`
-
+4. middleware trustedhost: wildcards for domain
 
 # v0.2.2
 1. middleware cors: support `allow_origin_regex`