+++ changelog.md
@@ -4,6 +4,7 @@
 3. lifespan: event_type support `shutdown`
 4. middleware trustedhost: wildcards for domain
 5. add middlewares: session , basehttp
+6. support: clear session
 
 # v0.2.2
 1. middleware cors: support `allow_origin_regex`